add `Formula`, creation via `~` like in R as macro / template magic
We can parse (or let the Nim parser) a formula starting with `~` and
store it in a `FormulaNode`, which we can serialize at will to perform
the calculation stored in it.

The syntax works just like in R (at least it should). Note however,
that the arguments to the formula *MUST* be untyped
identifiers. Support for variable injection etc. is not given. In that
case use the (not yet implemented) string or macro (using `{}`)
version of this.

I guess...