add freqPoly impl, handle drawing of multiple geoms, ...
Also adds option to set geom specific aesthetics.

We now create a new viewport covering the whole plot for each geom we
want to draw. This is added to the plot viewport as a child. By doing
this we can keep the order in which the user calls the `geom_*`
functions alive. Otherwise a histogram would *always* be drawn after a
line or points, simply because the histogram is implemented as
children viewports all along, whereas points draw on the viewport directly.