add the `f{}` macro to create FormulaNodes, including functions
The macro looks like this:
f{ "newValue" ~ "hwy" / "cty" * "10.0" }
It's basically strings separated by mathematical operators. However,
it also supports to use functions defined in the scope, e.g.:
f{ "normalized" ~ "hwy" / mean("hwy") }
where the `mean` function is understood to be a function of signature:
proc (s: PersistentVector[Value]): Value.

The `Value` type is added to have the ability to properly store
different datatypes in a dataframe (which will be added in a later
commit). The function mentioned before will then act on a whole column
"hwy" (in this case) and the value is replaced.