add validation functions for Value, FormulaNode combinations
The `isValidRow` assumes that `Value` is of kind VObject and contains
the key / value pairs corresponding to a single row of a data
frame. The proc then checks whether the given row is valid under the
given formula, e.g. if a given value of a key matches a predicate.

`isValidVal` on the other hand, checks whether a single `Value` (*not*
VObject) whether this valid mathces the RHS of the given formula node.