add DataFrame based on `OrderedTable[string, PersistentVector[Value]]`
This is a pretty naive implementation of a column based data frame
using Clojure's PersistentVector implemented by @PMunch.
The original idea was to be able to "copy" around the columns under
different mutations of the data frame, without performing a lot of
copying.

However, the way the code is implemented now does not really make a
whole lot of use of it anymore.
I still leave it in for now, because on a simple benchmark this
implementation was still ~10% faster than replacing the vector by a
normal `seq`.

The change from one to the other is a matter of a few lines of
code. Might be changed in the near future.