properly handle serialization of `FormulaNode` of kind fkFunction
The `fkFunction` serialization can only be used with the
`DataFrame` (or in principle any type that uses `Value` as its
internal data storage for each element).

Since the functions so far are all working on a whole column, the
result is the same for each index `idx` we call serialization with. So
we make the `FormulaNode` itself mutable and store the result of the
computation in the formula. Thus we can avoid having to calcualte
e.g. the `mean` or whatever thousands of times.