add pretty brute force `outerJoin` implementation
We first calculate the intersecttion of the key by which we join. Then
we get all indices for each data frame for which we find elements in
the intersection (a data frame may have duplicates).
Finally we walk over the indices and check from which data frame we
have to read elements and add step wise to the resulting data frame in
a while loop. This can hopefully be done in some nicer way.