big refactor of how `Scale` works, support for `size` aes
A `Scale` now is attached to each `Aesthetitcs`. Or rather potentially
more than one. A `Scale` is now more accurately represents the mapping
of some data to some scaling, i.e. mapping data to colors etc.

In the future `Scale` is now ready to also map things like `x`, `y`
data to `log` space etc.

This commit also adds things like guessing the type of a `seq[Value]`
column of a data frame.

Also support for `Scale.size` called via `aes(size = "something")` is
added. Although this only works for `geom_point` so far (we just need
to add an `aes` argument to the other geoms. Of course it's
potentially non trivial to handle those scales for other
geoms. E.g. in case of histograms a scaling via `color` actually also
means a grouping, since the histogram must be done for each label of
the scale individually.

Initial support for the distinction of continuous and discrete data is
also added. Although continuous data scales are not really quite
implemented, we can now distinguish it by looking at a column of a
data frame and guessing it.

Also the `GgPlot` object does not have any `Scale` anymore.

In a few places this still means that we hardcode the `GgPlot.aes`
instead of looking at `Geom.aes` (e.g. in histogram).

Also if several `Scales` in addition to `x` and `y` are used, their
legends will overlap currently.