make use of `enumerateScales` for `createPointGobj`
Cleans up the proc and more importantly we now do not read all of the
`Scale` data for each element that we want to plot, but rather for
each label. This is currently implemented by creating a filtered data
frame for each label and reading from that.

TODO: This breaks plotting against a non existing label as an
individual group!