add lenient math operators for `Value`, converts to VFloat
Adds math operators for `Value `type. Note that this only works for
VInt and VFloat. Also note that a `VInt` will be converted to `VFloat`
for the operation.
TODO: Differentiate between +,- and *,/ math procs, since for +,- we
can keep VInt a VInt, as long as both `v`, and `w` are VInts.