have `evaluate` w/ index return `Value`, supports `VBool` now
This means that `evaluate` working on an index can now also work with
`VBool` columns. Also there is less back and forth between different
data types.

Working with explicit string values using formulae is still up in the
air though, afaict, since we map `VString` to a DF column. We should
probably check whether a string actually corrresponds to a column and
if not, just use the literal string?