add `bind_rows` proc to stack two data frames
`bind_rows` stacks two data frames on top of one another. If they do
not share all columns, the rows of the non existant columns will be
set to `VNull`.

The resulting data frame has an `id` column, which contains a string
referencing which data frame the row corresponds to.