allow `evaluate` to work with VString values
We first check whether a `VString` corresponds to a column name. If it
does, we use the column at index `idx`. If it's not a column, we use
the literal value.

NOTE: This is of course problematic if the dataframe contains fields
that have the same value as a column name!