[formula] fix keys we yield in `groups` iterator
We accidentally yielded the current keys, which by the time we
actually yield something already point to the keys of the next group
we will yield.

Note that for the very last subgroup yielding `currentKeys` is fine,
since in that case `currentKeys == lastKeys`. If not we'd have yielded
a group before that!