[formula] fix comparison of two Value, fixes #1
NOTE: we still need to add a few more tests cases to this.

The reason it previously failed was that the `Value` in the data frame
was of type `VFloat`, while the Value we compared with was
`VInt`. Since our `==` was checking for exact equality (including
type), this comparison failed.

The mini test case in the `testsDf.nim` does cover that case, but we
should expand on that. Also add a case where it's both float / float and
int / int.