add readXYcols respecting geom x,y over ggplot x,y, fixes #4, add test
The `readXYcols` introduced here takes care of extracting the corrrect
x,y columns from a given ggplot and geom object. Then the data is read
into the given data type.

Tests are added, which try to check the behavior, but rely on the fact
that the given proc must actually be called and not be worked around.