[bind_rows] do not add to `id` column twice, TODO add test!
If there was a column of `id` already in one of the datasets that are
given to `bind_rows` we'd get a wrong size before. Have to write a
test that calls `bind_rows` multiple times in a row, where one of the
inputs is always the previous result of `bind_rows`. E.g. the
following
df1:
from    x   y
one     1   2
...

merged with
df2:
x    y
34   3
bind_rows([df1, df2], id = "from")
would have failed before.