allow handing width / ignore indices in addHistoRects
Handing a width in relative coordinates of the individual bar viewport
allows either for histograms non touching bars (if user desires; not
yet implemented) and more importantly sets base for bar plots, for
discrete data in which we do not want to have touching bars.

Ignoring specific indices on the other hand is also used for discrete
data so that we can define N + 2 viewports (via layout) and have the
outer most ignored, so that we can have some margin around the bars.