add createBarGobj
This is really WIP, because the tick positions for the labels don't
match up with the real positions of the bars. This is because the tick
loctions are calculated based on the plot viewpoint (4th child of main
viewport) instead of the invididual children of the 4th, in which the
bars are drawn. This is off, because we added 2 "dead" specific width
viewports for margins of the bar plot. Have to take that margin width
into account when calculating the tick positions of the locations.

Easy fix, but will have to wait :)