remove assertion, which doesn't hold w/ current ginger
The assertion implemented here holds for a modified version of ginger,
which treats the definition of `wView` and `hView` slightly
differently. In that case those would not correspond to the sizes of
the view of the parent (as currently) in pixels, but rather the size
of the current viewport.

This change results in a few breaking changes and I'm not certain it's
the desired behavior, but it was used for the implementation of the
bar centering fix at first. I leave the comment in as a note to myself.