allow `lift` templates to not export result, export `string` lifter
The previous definition of the lift templates was problematic, if one
wanted to lift some proc only locally (which may be necessary in some
circumstances, e.g. when lifting a proc for a unit test). Set the
`toExport` static bool argument to false.

Also exports the string templates for the user to use.