add custom `isDigit` template, fixes #20, remove unused funcs
Since the previously used `isDigit` proc from strutils was deprecated
and finally removed in https://github.com/nim-lang/Nim/pull/12535 we
have to use our own definition of a digit.

For now a normalized string (removes potential underscores) and then
checking if all chars are in {'0'..'9'} seems reasonable for our
purposes. This version now however is just called `isInt`. Note
thought that this is not defintely an int, but just likely (underscore
at beginnign and end are technically allowed by our definition).

Also removes some dead procs, which aren't used and are pretty
useless (a few of those still in the code though..).