make `f{}` work with args of field access, procs in symbols
Now `f{}` works both for cases where a called function is a field of
an object and also if simply a function is available under an
identifier and not a raw proc. This allows such a formula:

```nim
f{ s.col ~ s.trans(s.col) }
```

to compile (and work!), mapping the `s.col` to its transformed.