add `FilledGeom`, modifiy `FilledScales`
These two objects form the foundation of the refactor, which attempts
to better decouple the calculations required for a plot and the
drawing side.

With these the `generatePlot` and `createGobj...` procs only see a
`FilledScales` object from which only has to be read. No additional
calculations need to happen in the drawing procs.