rewrite scale filling / data access during plotting, also fixes #19
This commit performs a major rewrite of the way the data is prepared
for the plots and the data is accessed during the drawing process.

Instead of performing calculations when the drawing (that is
construction of GraphObjects) is done and having to modify different
parameters like scales etc. again and again is now completely avoided.

Everything required is now calculated beforehand and stored in a
`FilledScales` object, which has several `FilledGeom` objects (one for
each geom call) which bring everything that is required and provide a
unified, DataFrame based, access to the data required for each type of
geom.

This implementation might be somewhat slower for certain types of
plots, but the separation of drawing and preprocessing is worth it in
my opinion.

It might be slower simply due to more DataFrame based operations
taking place (due to inefficiencies in the DF impls) and possibly
unnecessary copies.
The former has now a direct motivation to be improved upon and the
latter will be fixed (which is now easier done).

There's still quite some code that is not needed anymore, which has to
be removed. With this commit simply all plots I created and used
before hand are now working correctly again and expands some features,
including adding (working) `stats` fields to the different geoms, so
that e.g. a geom_histogram can take `stat="identity"` and be drawn
from already precomputed bins and counts. Thus fixes issue #19.