return `cont` scales instead of col names, needed for cont styles!
The continuous scales are of course needed and not only the column
names referring to those scales, since the scales are needed to
retrieve the correct style information for the continuous scales! The
code for that was simply missing.

And yes, the `mapData` field of the `Scale` is of course still needed,
exactly for this.