specify "data" column when building FilledGeom & explain reason
The previous "brainfart" wasn't that smelly actually. It simply wasn't
explained so that it confused even me who wrote it.

The `data` DF being built in `fillScale` is only used for label
building / dataScale calculation and requires all data from all
columns, which are considererd of the same `aes` field, i.e. are
"compatible" (according to the user!). Thus we need all those in one
DF.