add reversed enumerateData iterator
It's essentially the same as `enumerateData` but returns the elements
in reveresed order. This is useful to construct stacked plots in the
correct order in each bin, that is the key with the highest string
value, e.g.g "suv" will be at the bottom and not at the top as would
normally happen, due to sorting of the DF before beign returned by the
`groups` iterator.