add `BinPositionKind`, move other bin rel. fields to statKind case
The binning information is not actually related to the geom at all,
but rather to the way he data is being handled. If we're supposed to
calculate the binning with the `histogram` proc we need this
information regardless of the geom.

Note that the order of application if all are set is:
1. binEdges
2. binWidth
3. numBins

BinPositionKind is an extension that doesn't exist (as far as I'm
aware, confused as I am) beyond ggplot2. It allows to set the
interpretation of the binning information of the `bins` (that's the
bin edges) column in the final plotting of the data. For non bin like
data it's simply `bkNone`, which means the geom is drawn exactly
at (x, y).

For bin like data it can be `bpLeft`, which means the geom starts at
`(x, y)`, `bpCenter` or `bpRight`. The latter two are mainly for geom
point and geom line so that a geom_point can be drawn into the center
of a bin instead of the left edge! Same goes for geom line (and geom_freqpoly).