fix bad numBins assertion and assignment to geom.numBins
The assertion that the resulting `numX` should be the same as the
input geom.numBins is bad due to the additional ways to set the number
of bins beyond just the `numBins` geom field (binWidth, breaks etc).

Then we also don't want to assign the `numX` back to the geom, since
that is supposed to store the value the user assigned
originally. Later we're going to use `numX` only anyways.