completely rewrite drawing logic
Instead of simply having one proc for each geom which has to deal with
all sorts of different cases we now divide the logic rather by:

- discreteness
- stat kinds
- and only then geoms

This allows to handle all sorts of different options in a more
streamlined way, especially regarding adding more options in the
future (at least I hope so).

Thought the code could probably still be compactified some more, but
well. Fine of for now.