allow VNull values when converting to float, represented as 0
This is useful for the cases when we have e.g. one additional bin edge
in our data, which we want to ignore.

NOTE: Another way to deal with this would be to explicitly remove the
right most edge from the DF before we build the `yieldDf` in the fill* procs.