add `dropNulls` option to `gather`
When gathering similar data from several columns one often encounters
the problems that while the data is of similar nature, it's not
actually the same number of elements. In this case the data may be
represented by several VNull values in the shorter column. To get rid
of those we can use the `dropNulls` argument to `gather` to remove
them while gathering.

NOTE: we should investigate this impl, since it's probably even less
efficient than the rest of the code, haha.