improve output of failed tests in `tCompareRecipes`
Instead of relying on `unittest.check`, which dumps the whole
diff (huge for our binary comparison) to the terminal, we compare
first, check if the comparison is true and if not we simply write
which file failed. Much more useful, since we have to compare by eye
anyways to see what went wrong.