add `readCsv` version working w/ streams, improve header handling
Sometimes one already has an existing stream of CSV/TSV like data. In
those cases a version of `readCsv` using streams directly is
useful. Also avoids having to creata a temp CSV file in the tests.

The handling of headers is improved. Although this comes with a
breaking change, if the header used is `#`.