add `unique` (cf dplyr.distinct) to remove duplicates from DF
The name is `unique` in Nim, because `distinct` is a keyword.

Either using all columns or only specific columns this allows to
remove rows, which are duplicates.

NOTE: If you select specific columns the first element matching
the given columns will be the one ending up in the final DF in the
current order of the DF!