[scales] only merge ginger.scales, if primary is set
If the primary scale `s1` in `mergeScales` is not yet set at
all, do not apply the typical min(low, low), max(high, high)
comparison, but first of all copy over the secondary scale.

This avoids setting the minimum to 0 in essentailly every case where
`s1` is still not set, e.g. visible in histogram plots.