avoid emphasizing string numbers in DF keys
When echoing a data frame, we annotate a number explicitly with `""`,
if it is stored as a string. This helps the user to see whether values
in a column were possibly wrongly parsed as strings instead of
numbers (Note that we can't simply annotate each column, since a
column can contain different types).

However, so far we also applied the same to the keys of a DF. This has
the nasty effect that accessing the columns cannot be done via
`df[$<someInt]`, but instead has to be `df["\"" & $<someInt> & "\""]`,
which is both unwieldly and not necessary, because keys are strings anyways.