[csv] improve handling of headers when using custom col names
If the CSV like input has a header with an explicit character at the
beginning of the line and custom columns are used, the header is
automatically skipped. If there is a header without an explicit
character and custom column names are used, the user has to provide
the number of lines of the header via `skipLines` to properly skip
it.
This way the more common case of assigning custom names to a file
without any header is more convenient.