add option to manually set discreteness of an axis, via scale_*_cont
In some cases the determination of discreteness for a column is not
optimal. In other cases the user simply wants to pretend a column is
discrete / continuous even though it isn't.

The current implementation uses an additional field `hasDiscreteness`
to differentiate between a custom discrete value and the `dcKind`
field. The problem is that we can't make `dcKind` an `Option`, because
it also acts as a variant discriminant.

We might be able to find a nicer solution to this however.

It is set by using the `scale_*_continuous` procs.