read discrete data according to x column
Instead of just reading data from the y column, we should sort by the
discrete scale `x` (TODO: handling of x/y interchange!) to actually
have the data in the correct order.
Note that this is mainly relevant for the order not of an additional
discrete scale e.g. via color etc. but rather if the data scale is
actually continuous.