apply trafo when filling cont. trans. scale, bound ticks in scale
Previously we didn't apply the scale transformation to the data
scale. This lead to a data scale that was in normal space. It's way
easier to handle if the data scale is already in the transformed
scale, since we apply a data transformation to the relevant DF columns
later and we compare those to the scale when determining if to
e.g. clip or drop data points.