introduce `GgStyle` w/ Option[T] to diff. user set / map / default
The `Option[T]` approach allows us to differentiate between user
settings (highest priority), user mappings (medium priority) and a
default style (as a fallback).

The only downside is that the user facing setting arguments now also
take `Option[T]` since we cannot define default values without
shadowing any possible user styles, e.g. say we had a `color`
argument:
`color = <anyColor>`
We could pick any color as the default and then check if the given
value is unequal to `<anyColor>` and only use the given value in that
case. But in general for all `GgStyle` types there is ambiguity. What
if the user wants to set `<anyColor>` specifically? For sizes we could
use negative values, but for enums and colors this is hard. :/