[formula] add inplace variants of select, mutate, transmute, simplify
Also simplifies the implementation by combining where possible.

The inplace variants are actually called =*inplace=. This is because
their use is somewhat discouraged and to not introduce a breaking
change (we cannot have mutable and non mutable have the same name!).

Performance *should* not be significantly worse for the non inplace
version, since the columns are PersistentVectors where each "copy" of
should just be a change of a reference.