[style] add `alpha` field to GgStyle to set alpha of `fillColor`
The `alpha` is a simpler way to access the alpha of a fill color. This
does explicitly not set the alpha of `color`.

If that is desired, we might want to add a new field for this as
well. The advantage of having `alpha` its own field is that we can
classify by fill color but then *only* change the alpha of the result.

If we were to also apply this to normal `color` e.g. applying an
`alpha` to a classified (by `fill`) `geom_freqPoly` we would also
change the alpha of the lines, which is not desirable in general.