make cont. scale `mapData` take DF instead of indices
This allows us to hand the DF at a later time and not capture it
already when filling the scale.

For combinations of a discrete and continuous scale (e.g. size and
color) we would get bad data beforehand, since the internal captured
DF contained all data instead of only the discrete, grouped parts to
determine the e.g. colors.