`handleDiscreteTicks` takes labels, handleTicks hideLabels, numTicks
Instead of handing a full scale, only hand the labels to allow calling
the proc from plaes where we don't have acces to a full scale.

handleTicks now allows to hide labels and set the number of ticks to
something other than the plot says (to call it from elsewhere,
e.g. for gradients).