add ridgeline plots
Done by adding `Ridges` object similar to done for `Facet` plots. The
idea is the same too. We create a layout for the ridges and draw in
each viewport a "full" plot. A little bit smarter than old facet,
because we have to only plot those labels, which correspond to the
ridge. This is the reason for the introduction of the =StyleLabel=