[column] allow conversion of int/float to string when reading
Convesion from float to string is done via `Value` so that we have one
function, which performs string conversion for floats.

We can change this in the future, but for now its
convenient. Converting to string is slow anyways and usually only used
for annotation etc and not huge amounts of data.