[postprocess] add note about assumption in `determineDataScale`
Only due to bugs in the code should we arrive at `determineDataScale`
without `s.col` being a well defined column in the DF.

NOTE: For constant values we might consider to change this behavior in
the future as to avoid wasting the space for such columns. Instead
when reading the columns in ggplot_drawing, we'd have to evaluate on
the DF, instead of just accessing the DF at index idx.