`isColumn` just checks whether stringification is column
This is valid, because for all formulas, which don't refer to existing
columns, we name the newly created columns by the stringification.

NOTE: this assumes that formulas using the LHS are not allowed,
because then this assumption would fail, since the correct name would
be fn.lhs / fn.colName!
But for `aes` (for which this is relevant) this is not really sensible
anyways, as far as I can tell right now in my sleep deprived state. :)