[df] add `clone` for df, low level asgn
The `[]=` should be a user facing proc, which should be reasonalby
safe. This means it should check the lengths (at least in debug mode)
and assign the length of the given vector as the DF length (for the
case of an empty DF).

This means we error out for ragged DFs.

In the future we should either raise an exception or just append Null
values in case a ragged DF is being created. For now keep it simple.