[postprocess] make sure we don't remove information from DF
We now have a dedicated function for the stringification of formulas
in the context of ggplotnim post processing. Since we have DFs with
reference semantics now and work inplace to apply transformations, we
would end up removing data if a transformation (specifically log10)
applies to the same column key.

Thus we assign to a different one so that we don't have to clone the
DF, which would introduce a significant performance overhead.