do not use Tensor pairs iterator
If we have to access the indices of a tensor, while iterating,
we *have* to avoid using the pairs iterator. Since the returned index
is a sequence of indices, we have to access it. That adds significant
overhead. In case of a simple filter this caused a performance penalty
of a factor of almost 10x for a DF w/ size 1e7.