`summarize` make sure to only add keys/labels once!
This is required, but a more important optimization would be to not
use the `groups` iterator for each formula!

But it's not important right now, because using the groups iterator is
extremely wasteful anyways, which is why we're rewriting the code to
calculate the code in a single pass over the DF on each groups at the
same time.

(factor 50x slower on DFs w/ 1e7 elements!)