`handleTicks` now properly respects hideTickLabels & margins
Instead of not even creating ticks and labels for `hideTickLabels =
true`, we do still create the ticks and labels now. This is useful,
because if we want to draw grid lines we need to know where the
ticks *would* be if they existed.

Also custom margins are not supported for ticks.