avoid `GgPlot` as generic
Originally it was envisioned that `GgPlot` could take different types
of data. However, by now the whole code base has become so entwined
with the data frame implementation that this is out of scope. There
isn't a good reason for it anyways and I'm pretty sure no one will
have used that.