add one-pass parsing of ints/floats based on @c-blake's work
This uses @c-blake's code from
https://github.com/nim-lang/Nim/pull/16055
to avoid having to copy the string which we parse and instead directly
parses a number into either an int or a string.

Note that this implementation has some edge cases that break it. In
particular a number that starts valid, but doesn't end as a valid
number (which is technically ok, but will break our CSV parser, since
our position in the memfile will *not* be on the next separator!).