[tests] only compare text related positions on linux
It appears that cairo on different platforms produces slightly
different font rendering. That means comparing to hardcoded values of
the position (if determined e.g. from strWidth).

This will still produce problems for things which are not text but may
be placed according to a string width. :(

I'm not sure if we do that anywhere though (e.g. tick labels are
placed according to strWidth of something else, but they are text themself).