do not import vega module by default (it needs webview)
Webview defines global variables, which means that if we import
vega_utils (now ggplot_vega) we always end up polluting our binaries
with webview related stuff even if we do not use vega.
That's not acceptable and breaks Nim CI.

This requires moving some stuff from `ggplotnim.nim` to the respective
submodules (which is good anyway).