add foundation to support arbitrarily scaled secondary axes
The code using `transFn` and inverse isn't working properly right now
for most transformations, because the current `scTransformedData`
code still more or less assumes that the transformation is "log
like" (i.e. powers of 10 are a desired tick label).

The tick computation for transformed scales has to change to support
this. Alternatively this code could be done such that transformed data
itself is treated as linear data. A transformation might be linear
after all.