fix previous value access for histogram drawing, performance fix
The previous code had the nasty side effect of creating a full tensor
for each individual element if the input was a constant
column (i.e. the case of no mapping of any kind, so no previous
values).

For very large inputs this caused *SEVERE* performance problems.

Instead we just get the column and access the idx'th element
directly. For a constant column this just means take the constant element.