add support for custom/customized color scales
Note: the added `ColorScale` type uses `uint32` colors, partially for
simplicity and partially because `geom_raster` is the most performance
sensitive color related plot we might create. As we need to create a
`seq[uint32]` for the raster drawing callback, it makes sense to
choose a representation such that further conversions are not
necessary there.