make size of geoms adjustable, `scale_size_discrete/continuous`
This finally makes it possible to adjust the range in which a
discrete (and in another commit continuous) size scale is set. The
default is 2 - 7 points, and now it can be set via the
`scale_size_discrete/continuous` procedures, which may be added to the
call chain.

Note: the continuous implementation isn't part of this commit, as git
is too confused with another addition (of continuous alpha scale),
such that untangling the diff is too annoying.