add distinction of mapping / setting for scales from DF data
This finally adds support to not only automatically map data from some
DF column to a certain scale (size, color, ...), but to also use DF
data to explicitly *set* the values of a certain scale, i.e. use sizes
stored in a DF column or colors stored as a string / int.

Note: a big chunk of the functionality is still missing as of this
commit, in particular to automatically deduce these things etc.