replace usage of `Option` to set geom scales, allow string / int colors
This commit marks the biggest change for the user of the library in a
long time (in a positive way).

Previously, any argument that *sets* a scale of a
geom (e.g. specifying the `size` of points, changing the `color` of
lines etc.) had to be given as an `Option[T]`. This has always been a
bit annoying and useless. Why should the user have to be forced to
write `some(1.5)`? This has always been a limitation of our logic, to
make sure we can detect what is a default value vs. what is actually
given by the user. Essentially, not for all types a good default value
that may never be used by the user to *set* was unambiguously
possible.

This commit changes the whole logic to a multi generic approach:

We add a `Missing` type and each argument to all the `geom_*`
procedures become types of the form:

```nim
PossibleFloat = Missing | SomeNumber | string | Option[float]
...
```

What this means is the following:
By default each optional argument is set to `missing()` (which just
returns a `Missing` instance). This way we can deduce at CT whether an
argument was given or not (if it's type `Missing` it means use the
default).

This means we can now use any number (in case of float like arguments,
e.g. size, ...) as input and *know* the argument was actually given by
the user.

The `Option[float]` argument is part of the type class to make sure
old code continuous to work!

You may notice the weird possible "string" type as well. This is to
allow the (added in a previous commit) distinction between mapping
values based on data (i.e. an argument of in `aes`), setting to a
constant value (in case of `size` by giving a number) *or* setting,
but based on a DF column. If a `string` is given, it's interpreted as
referring to a column of the input dataframe. The values found in that
column will then be used explicitly to *set* the scale.

For `Color` scales (fill and color) we also go one step further and
now allow to set colors not only by providing `chroma.Color` types,
but also by using `int` values (interpreted as `uint32` colors) and
`string` based colors (by treating them as HTML valid color strings).

This means code as follows is now possible:

```nim
import ggplotnim

let df = toDf({ "x" : @[1, 2, 3, 4], "y" : @[1, 2, 3, 4],
                "colors" : @["red", "green", "blue", "#FF00FF"]})
ggplot(df, aes("x", "y")) +
  geom_point(aes = aes(color = "colors"), size = 12.3) +
  scale_color_identity() +
  ggsave("/tmp/colors_manual.pdf")
ggplot(df, aes("x", "y")) +
  geom_point(color = "colors", size = 12.3, alpha = 0.6) +
  ggsave("/tmp/colors_automatic.pdf")
```

Note the usage of referring to a column in the `color` argument,
either as an `aes` (as usual) and then telling it to treat the scale
`color` as *the identity*, or by just explicitly *setting* a color to
a string ↦ interpreted as a column, because "colors" is not a valid
HTML color itself.
In addition we see the usage of explicitly setting sizes & alphas
based on pure numbers instead of `Option[float]`.