add PossibleSecondaryAxis to clean up sec_axis logic
The reason for adding this is that due to the Datamancer change that
`FormulaNode` is now `Formula[Column]` with the default `sec_axis =
sec_axis()` logic, we'd get a "cannot instantiate generic" error. This
fixes it and cleans up code.

Also adds a missing =scale_y_discrete= that takes labels.