add TODOs about making legend fonts customizable
At the same time we should probably also allow to customize the
actual *text* written over a legend! Ideally this is done in a rewrite
in which we generalize the legend construction from a fixed kind of
input data and add a procedure to generate a custom legend based on
given elements, texts etc.