[legend] let legend size overflow for arbitrary sizing
The overflow check for the legend is not super useful. If one wishes
to move the legend position to somewhere else (e.g. in a facet plot
with 3 plots there's empty space at the bottom right), it makes sense
to allow the plot range to fill the full plot, which depending on
inputs might cause an overflow in the layout logic.