fix code using direct formulas combined with `log10` (or any transf)
Code like this:

```nim
ggplot(df, aes("x", "y")) +
  geom_point() +
  # define errors as a formula, which references our "y" scale
  geom_errorbar(aes(yMin = f{`y` - 0.1}, yMax = f{`y` + 0.1})) +
  scale_x_log10() + scale_y_log10() +
  ggsave("/tmp/ErrorBar.pdf")
```

was previously broken, because internally we rewrote the formula for
yMin / yMax to take the string representation of the argument at face
value and tried to apply the transformation to a column of that name,
which didn't actually exist yet. Now we make sure to create such a
column first.