fix issue with drawing transformed error bars and other plots
We were accidentally overwriting a transformed column by its non
transformed column in the previous code. This fixes that. However, I
did not have the time to verify that this does not break something
else. Because maybe I used `c.col.evaluate` for a reason in the
past. The tests seem to pass though.